chore: rename video hash files
